Move Docsify wiring to adapters/docsify/index.js — IoC refactor
index.js is now a pure core re-export (no DOM, no Docsify).
All $docsify.plugins + hook.* logic lives in adapters/docsify/index.js.
Fix: preserve node.position in yamlComponents so serializeToMarkdown
can apply string replacements correctly.
renderer.js removed (orphaned after index.js took over).
