+++ docs/index.html
@@ -105,8 +105,8 @@
   <!-- Peer dep: js-yaml must be a global before the plugin module runs -->
   <script src="https://cdn.jsdelivr.net/npm/js-yaml@4/dist/js-yaml.min.js"></script>
 
-  <!-- THE plugin — everything else is wired inside -->
-  <script type="module" src="/packages/docsify-plugin/src/index.js"></script>
+  <!-- Docsify adapter — wires core engine to Docsify lifecycle -->
+  <script type="module" src="/packages/docsify-plugin/src/adapters/docsify/index.js"></script>
 
   <!-- Docsify core + search (deferred — plugin module registers hooks first) -->
   <script defer src="https://cdn.jsdelivr.net/npm/docsify@4"></script>
